WIOS-82 Fixed analyzer warnings for Xcode 5.1, LLVM 5.1

+++ Classes/Keychain/DUKeychain.m
@@ -119,7 +119,7 @@
 		
 	OSStatus status = SecItemDelete((__bridge CFDictionaryRef)query);
 	if (status != errSecSuccess) {
-		DULog(@"SecItemDelete failed: %ld", status);
+		DULog(@"SecItemDelete failed: %d", (int)status);
 	}
 }
