+++ Classes/NS+Extensions/NSMutableArray+Extensions.m
@@ -15,14 +15,14 @@
     if(!seeded)
     {
         seeded = YES;
-        srandom(time(NULL));
+        srandom((int)time(NULL));
     }
     
     NSUInteger count = [self count];
     for (NSUInteger i = 0; i < count; ++i) {
         // Select a random element between i and end of array to swap with.
-        int nElements = count - i;
-        int n = (random() % nElements) + i;
+        NSInteger nElements = count - i;
+        NSInteger n = (random() % nElements) + i;
         [self exchangeObjectAtIndex:i withObjectAtIndex:n];
     }
